+++ Database_Project_Group1/src/Song.java
@@ -2,7 +2,6 @@ import java.sql.Date;
 import java.util.ArrayList;
 
 public class Song {
-
     public int songId;
     public String songName;
     public int songDuration;
@@ -68,7 +67,7 @@ public class Song {
     }
 
     /**
-     * selectAll method of the Song class.
+     * selectSong method of the Song class.
      * This method will use the getData(String, ArrayList<String>) method of the Alinity
      * class. The SQL statement will select Songs where the ID value input will be bound
      * from the ArrayList, and save the result.
@@ -79,17 +78,22 @@ public class Song {
      *
      * @throws AlinityException
      */
-    public void selectAll(ArrayList<String> stringList) throws AlinityException {
+    public void selectSong(User user, String songName) throws AlinityException {
         try {
-            ArrayList<ArrayList<String>> result = AlinityMain.alinityDB.getData("SELECT * FROM Song WHERE songId = ?" , stringList);
-            System.out.print("\nColumn headers: " + result.get(0));
-            ArrayList<String> songData = result.get(1);
-            setSongId(Integer.parseInt(songData.get(0)));
-            setSongName(songData.get(1));
-            setSongDuration(Integer.parseInt(songData.get(2)));
-            setAlbumId(Integer.parseInt(songData.get(3)));
-            setArtistId(Integer.parseInt(songData.get(4)));
-            setGenreId(Integer.parseInt(songData.get(5)));
+            if(user.getRole().equals("General") || user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(songName);
+                ArrayList<ArrayList<String>> result = AlinityMain.alinityDB.getData("SELECT * FROM Song WHERE songId = ?", info);
+                System.out.print("\nColumn headers: " + result.get(0));
+                ArrayList<String> songData = result.get(1);
+                setSongId(Integer.parseInt(songData.get(0)));
+                setSongName(songData.get(1));
+                setSongDuration(Integer.parseInt(songData.get(2)));
+                setAlbumId(Integer.parseInt(songData.get(3)));
+                setArtistId(Integer.parseInt(songData.get(4)));
+                setGenreId(Integer.parseInt(songData.get(5)));
+                printSong();
+            } else System.out.println("You do not have access to this function. Please contact an administrator.");
         } catch (IndexOutOfBoundsException ioobe) {
             throw new AlinityException(ioobe, "-> Error in obtaining data (IndexOutOfBoundsException) from the database. Please check your syntax in the selectAll(ArrayList<String>) method.", "SELECT * FROM Song WHERE songId = ?");
         } catch (NullPointerException npe) {
@@ -98,7 +102,7 @@ public class Song {
     }
 
     /**
-     * updateAll method of the Song class.
+     * updateSong method of the Song class.
      * A SQL statement as a String is created using the current name, duration,
      * albumId, artistId, and genreId where is based on the values which will be bound
      * from the input ArrayList of Strings.
@@ -109,10 +113,18 @@ public class Song {
      *
      * @throws AlinityException
      */
-    public boolean updateAll(ArrayList<String> stringList) throws AlinityException {
+    public boolean updateSong(User user, String songName, String songDuration, Album album, Artist artist, Genre genre) throws AlinityException {
         try {
-            String putStmt = "UPDATE Song SET songName = ? , songDuration = ?, albumId = ?, artistId = ?, genreId = ? WHERE songId = ?";
-            return AlinityMain.alinityDB.setData(putStmt, stringList);
+            if(user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(songName);
+                info.add(String.valueOf(songDuration));
+                info.add(String.valueOf(album.getAlbumId()));
+                info.add(String.valueOf(artist.getArtistId()));
+                info.add(String.valueOf(genre.getGenreId()));
+                String putStmt = "UPDATE Song SET songName = ? , songDuration = ?, albumId = ?, artistId = ?, genreId = ? WHERE songId = ?";
+                return AlinityMain.alinityDB.setData(putStmt, info);
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator."); return false;
         } catch (NullPointerException npe) {
             throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the updateAll(ArrayList<String>) method.", "UPDATE Song SET songName = ? , songDuration = ?, albumId = ?, artistId = ?, genreId = ? WHERE songId = ?");
         } catch (IndexOutOfBoundsException ioobe) {
@@ -121,7 +133,7 @@ public class Song {
     }
 
     /**
-     * insertAll method of the Song class.
+     * insertSong method of the Song class.
      * A SQL statement as a String is created, which takes in
      * values based on the values to be bound from the ArrayList of Strings.
      * Executes using the setData method.
@@ -130,19 +142,27 @@ public class Song {
      *
      * @throws AlinityException
      */
-    public boolean insertAll(ArrayList<String> stringList) throws AlinityException {
+    public boolean insertSong(User user, String songName, String songDuration, Album album, Artist artist, Genre genre) throws AlinityException {
         try {
-            String insertStmt = "INSERT INTO Song (songName, songDuration, albumId, artistId, genreId) VALUES (?, ?, ?, ?, ?)";
-            return AlinityMain.alinityDB.setData(insertStmt, stringList);
+            if(user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(songName);
+                info.add(String.valueOf(songDuration));
+                info.add(String.valueOf(album.getAlbumId()));
+                info.add(String.valueOf(artist.getArtistId()));
+                info.add(String.valueOf(genre.getGenreId()));
+                String insertStmt = "INSERT INTO Song (songName, songDuration, albumId, artistId, genreId) VALUES (?, ?, ?, ?, ?)";
+                return AlinityMain.alinityDB.setData(insertStmt, info);
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator."); return false;
         } catch (NullPointerException npe) {
-            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the insertAll(ArrayList<String>) method.","INSERT INTO Song (songName, songDuration, albumId, artistId, genreId) VALUES (?, ?, ?, ?, ?)");
+            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the insertSong(ArrayList<String>) method.","INSERT INTO Song (songName, songDuration, albumId, artistId, genreId) VALUES (?, ?, ?, ?, ?)");
         } catch (IndexOutOfBoundsException ioobe) {
-            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the insertAll(ArrayList<String>) method.","INSERT INTO Song (songName, songDuration, albumId, artistId, genreId) VALUES (?, ?, ?, ?, ?)");
+            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the insertSong(ArrayList<String>) method.","INSERT INTO Song (songName, songDuration, albumId, artistId, genreId) VALUES (?, ?, ?, ?, ?)");
         }
     }
 
     /**
-     * deleteAll method of the Song class.
+     * deleteSong method of the Song class.
      * Deletes the given song data where the songId is bound
      * via tha value inside the ArrayList of Strings.
      * Executes using the setData method.
@@ -152,24 +172,28 @@ public class Song {
      *
      * @throws AlinityException
      */
-    public boolean deleteAll(ArrayList<String> stringList) throws AlinityException {
+    public boolean deleteSong(User user) throws AlinityException {
         try {
-            String deleteStmt = "DELETE FROM Song WHERE songId = ?";
-            return AlinityMain.alinityDB.setData(deleteStmt, stringList);
+            if(user.getRole().equals("Admin")) {
+                ArrayList<String> info = new ArrayList<>();
+                info.add(String.valueOf(this.getSongId()));
+                String deleteStmt = "DELETE FROM Song WHERE songId = ?";
+                return AlinityMain.alinityDB.setData(deleteStmt, info);
+            } else System.out.println("You do not have the correct permissions to use this function. Please contact an administrator."); return false;
         } catch (NullPointerException npe) {
-            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the deleteAll() method.", "DELETE FROM Song WHERE songId = ?");
+            throw new AlinityException(npe, "-> Error in manipulating data (NullPointerException) from the database. Please check your syntax in the deleteSong() method.", "DELETE FROM Song WHERE songId = ?");
         } catch (IndexOutOfBoundsException ioobe) {
-            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the deleteAll() method.", "DELETE FROM Song WHERE songId = ?");
+            throw new AlinityException(ioobe, "-> Error in manipulating data (IndexOutOfBoundsException) from the database. Please check your syntax in the deleteSong() method.", "DELETE FROM Song WHERE songId = ?");
         }
     }
 
     /**
-     * printAlbum method of the Song class.
+     * printSong method of the Song class.
      * Print values for the current songName,
      * songDuration, and the id of the artist and genre (will be changed
      * later to represent the song itself).
      */
-    public void printAlbum() {
+    public void printSong() {
         System.out.println("\nID: " + this.songId);
         System.out.println("Song Name: " + this.songName);
         System.out.println("Song Duration: " + this.songDuration);
